Modernize C++ by replacing legacy patterns and fixing warnings.
- Replace `size() != 0`, `size() > 0` with `!empty()` in 14 files.
- Replace C-style casts `(int)`, `(float)`, `(uint8_t)` with `static_cast` in targeted rendering files.
- Fix `-Wreorder` warning in `source/rendering/core/graphics.cpp` (GameSprite::TemplateImage).
- Fix `-Wignored-qualifiers` warning in `source/game/complexitem.h`.
- Fix `-Wdeprecated-enum-enum-conversion` in `source/ui/toolbar/standard_toolbar.cpp`.
- Fix wxWidgets compatibility issue in `source/app/application.cpp` for wx 3.3+.
- Verified build with `./build_linux.sh`.

+++ source/ui/toolbar/standard_toolbar.cpp
@@ -41,7 +41,7 @@ StandardToolBar::StandardToolBar(wxWindow* parent) {
 	toolbar->AddTool(wxID_COPY, wxEmptyString, copy_bitmap, wxNullBitmap, wxITEM_NORMAL, "Copy (Ctrl+C)", wxEmptyString, nullptr);
 	toolbar->AddTool(wxID_PASTE, wxEmptyString, paste_bitmap, wxNullBitmap, wxITEM_NORMAL, "Paste (Ctrl+V)", wxEmptyString, nullptr);
 	toolbar->AddSeparator();
-	toolbar->AddTool(MAIN_FRAME_MENU + MenuBar::JUMP_TO_BRUSH, wxEmptyString, find_bitmap, wxNullBitmap, wxITEM_NORMAL, "Jump to Brush (J)", wxEmptyString, nullptr);
+	toolbar->AddTool(static_cast<int>(MAIN_FRAME_MENU) + static_cast<int>(MenuBar::JUMP_TO_BRUSH), wxEmptyString, find_bitmap, wxNullBitmap, wxITEM_NORMAL, "Jump to Brush (J)", wxEmptyString, nullptr);
 	toolbar->Realize();
 
 	toolbar->Bind(wxEVT_COMMAND_MENU_SELECTED, &StandardToolBar::OnButtonClick, this);
